# src/main/java/be/projetSGBD/api/AssociationApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package be.projetSGBD.api;

import be.projetSGBD.model.Association;
import java.math.BigDecimal;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@Validated
@Api(value = "Association", description = "the Association API")
public interface AssociationApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/v1/addAssociation : Create the association
     * Create the association
     *
     * @param association association object (optional)
     * @return created (status code 201)
     */
    @ApiOperation(value = "Create the association", nickname = "createAssociation", notes = "Create the association", tags={ "Association", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "created") })
    @RequestMapping(value = "/api/v1/addAssociation",
        consumes = { "application/xml", "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Void> createAssociation(@ApiParam(value = "association object"  )  @Valid @RequestBody(required = false) Association association) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/v1/association/{idAssociation} : Delete the association
     * Deletes the association
     *
     * @param idAssociation association id (required)
     * @return successful operation (status code 204)
     *         or Given user ID doesn&#39;t exist (status code 404)
     */
    @ApiOperation(value = "Delete the association", nickname = "deleteAssociation", notes = "Deletes the association", tags={ "Association", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "successful operation"),
        @ApiResponse(code = 404, message = "Given user ID doesn't exist") })
    @RequestMapping(value = "/api/v1/association/{idAssociation}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deleteAssociation(@ApiParam(value = "association id",required=true) @PathVariable("idAssociation") BigDecimal idAssociation) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/association/{idAssociation} : Returns the association
     * Returns the association
     *
     * @param idAssociation association Identifier (required)
     * @return successful operation (status code 200)
     *         or Given user ID doesn&#39;t exist (status code 404)
     */
    @ApiOperation(value = "Returns the association", nickname = "getAssociationByIdAssociation", notes = "Returns the association", tags={ "Association", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation"),
        @ApiResponse(code = 404, message = "Given user ID doesn't exist") })
    @RequestMapping(value = "/api/v1/association/{idAssociation}",
        method = RequestMethod.GET)
    default ResponseEntity<Void> getAssociationByIdAssociation(@ApiParam(value = "association Identifier",required=true) @PathVariable("idAssociation") Integer idAssociation) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}