# src/main/java/be/projetSGBD/api/PatientApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package be.projetSGBD.api;

import java.math.BigDecimal;
import be.projetSGBD.model.Patient;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@Validated
@Api(value = "Patient", description = "the Patient API")
public interface PatientApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/v1/addPatient : Create the patient
     * Create the patient
     *
     * @param patient user object (optional)
     * @return created (status code 201)
     */
    @ApiOperation(value = "Create the patient", nickname = "createPatient", notes = "Create the patient", tags={ "patient", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "created") })
    @RequestMapping(value = "/api/v1/addPatient",
        consumes = { "application/xml", "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Patient> createPatient(@ApiParam(value = "user object"  )  @Valid @RequestBody(required = false) Patient patient) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/v1/patient/{idPatient} : Delete the patient
     * Deletes the patient
     *
     * @param idPatient patient id (required)
     * @return successful operation (status code 204)
     *         or Given user ID doesn&#39;t exist (status code 404)
     */
    @ApiOperation(value = "Delete the patient", nickname = "deletePatient", notes = "Deletes the patient", tags={ "patient", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "successful operation"),
        @ApiResponse(code = 404, message = "Given user ID doesn't exist") })
    @RequestMapping(value = "/api/v1/patient/{idPatient}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deletePatient(@ApiParam(value = "patient id",required=true) @PathVariable("idPatient") Long idPatient) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/patient/{idPatient} : Returns the patient
     * Returns the patient
     *
     * @param idPatient Customer Identifier (required)
     * @return successful operation (status code 200)
     *         or Given user ID doesn&#39;t exist (status code 404)
     */
    @ApiOperation(value = "Returns the patient", nickname = "getPatientByIdPatient", notes = "Returns the patient", tags={ "patient", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation"),
        @ApiResponse(code = 404, message = "Given user ID doesn't exist") })
    @RequestMapping(value = "/api/v1/patient/{idPatient}",
        method = RequestMethod.GET)
    default ResponseEntity<Patient> getPatientByIdPatient(@ApiParam(value = "Customer Identifier",required=true) @PathVariable("idPatient") Integer idPatient) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}